+++ CHANGELOG.md
@@ -16,7 +16,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - BREAKING: Remove deprecated command `cluster instances migrate storage-version` previously used 
   to set new storage version for Astarte / AVI CRs (used for Astarte Operator v22.11 migration).
 - BREAKING: Remove deprecated operator management commands `astartectl cluster *-operator`
-
+- BREAKING: Remove deprecated instance deploy / show / destroy commands: `astartectl cluster instance show`,
+  `astartectl cluster instance deploy`, `astartectl cluster instance destroy`. Obsolete deployment
+  profiles are also removed.
+  
 ## [24.5.3] - 2026-01-28
 ### Changed
 - `appengine device attributes list`: display attributes in a human-friendly